requirements.txt
obspy
pymongo
matplotlib
numpy
schedule
python-telegram-bot 